# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Start API",
        "description": "AgroBozor uchun backend API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://daladan-api.loc/api/v1",
            "description": "Local dev server (OSPanel)"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "php artisan serve"
        }
    ],
    "paths": {
...
        "/resources/cities": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Faol tumanlar ro'yxati (ixtiyoriy region_id bo'yicha)",
                "description": "cities() — GET /resources/cities",
                "operationId": "0d41ec81256a3feecd7b0d99e59fabe4",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "Faqat shu viloyatga tegishli tumanlar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tumanlar ro'yxati"
                    }
                }
            }
        },
...
        "/resources/subcategories": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Faol subkategoriyalar ro'yxati (ixtiyoriy category_id bo'yicha)",
                "description": "subcategories() — GET /resources/subcategories",
                "operationId": "d29b942b23445d27362e2eb4911c49dd",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Faqat shu kategoriyaga tegishli subkategoriyalar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subkategoriyalar ro'yxati"
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "O'zi haqida ma'lumot",
                "description": "profile() — GET /profile",
                "operationId": "619297248f5a11bac7e6dc42cb107928",
                "responses": {
                    "200": {
                        "description": "Profil ma'lumotlari"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Profilni yangilash",
                "description": "updateProfile() — PUT /profile",
                "operationId": "798da08340d38b76e1f29e150721dfa4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fname": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "lname": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "telegram": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "region_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil yangilandi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/avatar": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Avatarni yangilash",
                "description": "updateAvatar() — POST /profile/avatar",
                "operationId": "6284eaffea00bdf2e1728fc3bc72d72e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar yangilandi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Parolni o'zgartirish",
                "description": "updatePassword() — PUT /profile/password",
                "operationId": "4369b5f7123569e7d63b6e5ff819fc01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "type": "string"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Parol muvaffaqiyatli o'zgartirildi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validatsiya xatosi / parol mos kelmadi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}